Rename AuthorizationHeader to Signature.
