+++ lib/awsraw.rb
@@ -1,8 +1,8 @@
 require "awsraw/version"
-require "awsraw/s3/authorization_header"
 require "awsraw/s3/canonicalized_resource"
 require "awsraw/s3/client"
 require "awsraw/s3/faraday_middleware"
 require "awsraw/s3/query_string_signer"
+require "awsraw/s3/signature"
 require "awsraw/s3/string_to_sign"
